requirements.txt
# Django
django>=5.0.1,<6.0.0 # 5.0.1 이상 6.0.0 미만

# DRF(Django Rest Framework)
djangorestframework>=3.14.0,<4.0.0

psycopg2>=2.9.0,<3.0

# drf-spectacular
drf-spectacular>=0.27.1,<0.28

# 4. channels
channels>=4.0.0,<4.0.1